Insist on interpreter short name in Python factor
We currently allow users to request a Python factor without an
interpreter short name prefix. This is highly unusual and rather error
prone, given it's likely to conflict with e.g. versions of other
packages. We shouldn't really support this so start logging a warning,
effectively deprecating this and allowing us to remove this
functionality later.

Signed-off-by: Stephen Finucane <stephen@that.guru>