Fix the type annotation for `color` param of `Spinner.finalize()`
Colorama's ANSI class attributes are defined as ints.
However, when an ANSI class is instantiated
(say, when the `AnsiFore` class is instantiated as `Fore`),
instance attributes with identical names are dynamically created,
but the attributes are strings containing ANSI escape sequences.

Thus, `AnsiFore.YELLOW` is an int, while `Fore.YELLOW` is a str.
mypy 0.991 isn't catching this discrepancy, but PyCharm did.

Fixes #2878