Remove unnecessary Future usage
The tox.config.api.Loader.load() method was manually creating a Future
and passing it to the .build() context manager method, but no executor
was involved and nothing was asynchronous.  This commit moves the call
to self.to() from the .load() method to the .build() method.  This
allows dropping the Future and changing .build() from a context manager
to a regular method.