Set the --parallel default to "auto", not CPU count
There is no functional reason to have the ArgumentParser default be
auto_detect_cpus(), as the option goes through a parser,
"parse_num_processes", that replaces "auto" with auto_detect_cpus(),
i.e. the CPU count.

Making the default "auto" makes the documentation clearer and
reproducible. Currently the tox docs say:
   -p VAL, --parallel VAL - run tox environments in parallel, the
   argument controls limit: all, auto - cpu count, some positive number,
   zero is turn off (default: 2)
That is misleading to end users: the default is not "2", but "auto"; 2
just happens to be the CPU count of the machine that generated these
docs.

Note that there was a DEFAULT_PARALLEL constant, that was however
unused, as a result of an earlier refactoring. Reuse that constant for
"auto" here.