docs(plugin): explain plugin registration
In the legacy branch, the plugin documentation described plugins get
discovered by being in the same environment. That documentation was not
ported to the new branch. The doc was originally added by
hff7143e7c991b1a80e7ec1ea6836ef3a21b5a812

I also wondered how the discoverability works via Pluggy, that is done
by having the plugin to register a `tox` entry-point which allows tox to
find the plugin module. Document that.

While writing a plugin, that would have helped me find out how to
install or enable it when it is indeed automatic (as long as a tox
entry-point is defined by the plugin).