docs(config): Fix base_python default reference
The previous description of `{package_root}` doesn't make much sense because it doesn't have anything to do with Python versions.  I'm guessing it was a copy and paste that wasn't updated.

Change it to a value that tries to be both short and accurate matching the conventions observed in other options.