Apply overrides from the prefix / generic testenv
An override to `testenv` is not applied to more specific testenv when
they define the same setting, even if one asks to append.

Given a `tox.ini`:

    [testenv]
    set_env = WHERE=top

    [testenv:test]
    set_env = WHERE=test

If one uses `tox -x testenv.set_env+=CI=1`, the `CI=1` is appended to
the generic testenv:

    $ tox config -k set_env
    set_env =
	WHERE=top
	CI=1

But is not seen in the more specific `testenv:test`:

    $ tox config -k set_env -e test
    set_env =
	WHERE=test

That is due to the override being set to `testenv` and never looked up
by `testenv:test`.

Change the load overides from its prefix section.

This let one forces appending values to list regardless of what is set
in the users `tox.ini`.

Ref: https://phabricator.wikimedia.org/T348871