docs/config.rst: State in config directive sections their ini file sections
The config file documentation is in essentially three parts: config files
and their sections, directives for the global config section, and
directives for the environment sections. For the second and third parts,
state again what config file sections they belong in.

(This is arguably redundant, but the page as a whole is long enough that
it's good to clarify this at the head of the directives in the not-unlikely
case that a new user didn't carefully peruse and understand the initial
section.)