tox.ini: diff-cover uses main@{upstream} instead of origin/main
These changes affect developers in the following way:
- Those with push access to the upstream rep (tox-dev/tox.git) see no
  difference.
- Those who fork and pay attention to which `main` branch they're using in
  their local repos have an increased chance of diff-cover working.
- Those who fork and don't pay attention to that may see an improvement or
  may see things break, but if it does break they have a bigger problem to
  fix anyway.

Rather than assuming that the remote `main` branch against which we wish to
diff is always on a remote called `origin`, make an (only somewhat) better
assumption that the remote branch which the local `main` is tracking is the
branch against which we wish to diff.

There are plenty of cases where this makes no difference or merely moves
some problems around, but this allows developers to use any remote names
they wish for the for the upstream and fork repos. (E.g., I call my remote
for the upstream that diff-cover _should_ be diffing against, `upstream`.)
The developer need only ensure that his local `main` branch is tracking the
correct remote. (The local main branch does _not_ have to be up to date
with it.) This is extremely easy to explain, just tell the developer to:

    #   `remote add` tox-dev/tox only if you don't have it.
    #   You may use any name for `UPSTREAM`.
    git remote add UPSTREAM https://gitub.com/tox-dev/tox.git
    git branch main --set-upstream-to=UPSTREAM/main

(That's probably worth documenting somewhere, but I don't know exactly the
best place for it.)

This is useful not just for diff-cover, but generally to make sure that the
developer is starting new dev branches from the latest `main`. (More on
this below.)

This commit does not have a changelog entry because the change is visible
only to developers who are running the tests.

*** More details, and the reasoning behind this change:

With the standard way that get PRs work, "outside" contributors will be
using two remote repos: the upstream in tox-dev/tox.git, which contains the
"true" `main` branch, and their own fork MYNAME/tox.git which contains
their development branch(s) that they're going to submit as a PR.

The `origin` remote might refer to either of these, depending on how the
developer originally cloned the local repo and what he decided to do when
adding the second remote. Particularly important is that in the fork repo
the `main` branch (if the developer chooses to keep it) will go far out of
date as time passes unless the developer makes a specific effort to fetch
changes on the upstream `main` and push them. Note that this will also
happen if the developer does not have a remote for the upstream; there's
not much we can do about this.

What diff-cover wants to be diffing against is the upstream `main` branch;
diffing against a fork's potentially ancient copy of that is not useful.
There are various ways the developer might have set up his development
repo, but using `main@{upstream}` should work as well as reasonably
possible for all the common ones.

1. Developer has push access to upstream; he's using only one remote repo
   and this works regardless of the remote name he uses.

2. Developer clones upstream and creates a fork after: the `main` branch
   will be the upstream main, and this will always use the latest version
   as of the most recent fetch.

3. Developer forks and clones the fork: the fork `main` branch will be up
   to date as of the time of the fork and will slowly age. This will be
   working against older upstream `main` code as this ages, but the
   developer too will not have newer changes to merge into his development
   branch as time passes. Whatever solution he uses to fix this (typically
   a manual update of fork `main` from upstream `main`) will also fix
   diff-coverage.