session/test_env_select: Move direct tests to the top of the file
All but one of the tests in this test module are actually testing
env_select functionality through calling other code that in turn
uses env_select (CliEnv, EnvSelector and register_env_select_flags).

We move that one test to the top of the file, but we do not indicate
the boundary between the groups (or even explain that the grouping
exists) at the request of @gaborbernat.