session.env_select.CliEnv: Document and test spaces behaviour (#3208)
It turns out that spaces on either side of environment names in a
comma-separated list are considered not part of the environment
name and removed. This was neither documented nor tested
directly; that's now fixed. (It was and is still also tested in
`config.cli.test_cli_ini.test_ini_exhaustive_parallel_values` as
well.)

The discovery of this was documented in issue #3207.