config.cli.test_cli_ini: Re-order for clarity (textual moves only)
No code changes; just a "move text" refactoring.

This makes the code quicker to understand by moving the exhaustive_ini
fixture next to the two tests using it, rather than having it separated
from them by several other test functions that don't use it. (This
separation is why I didn't immediately see an important point in that
in issue #3207.)

There's also a slight re-ordering of tests; they're still in an order
that makes sense for testing the functionality from the smallest to
the largest bits.