Consider complete envs, not factors
To determine whether we have been given a valid *env*, we compare the
*factors* of the *env* to our known set of *factors*. If a testenv
contains a series of valid factors, we assume the testenv is valid. This
is an incorrect assumption: a testenv is only valid if it defined in the
configuration file *or* is one of the "special" auto-generated testenvs
like e.g. 'py310'. Correct this mistake.

While we're here, we make use of verbose mode to explain our regex for
matching "magic" factors better.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes: #3219