Cache env config
This commit adds a new cache to the main config object, specifically for
env config. When env config is retrieved via Config.get_env(), it is
not in general possible to determine whether the env is a packaging
environment or not. Because retrieving standard section config needs to
know which "base" section to inherit from, and this is different for
test envs and packaging envs, we require a separate way to retrieve
previously created env config. At creation time we do reliably know
whether this is a packaging or run env, and we want that to persist (if
it changes, we call clear_env() and recreate it with the new type).