Support external tox.pytest usage via "test" extra (#3422)
In order to allow plugin writers to easily pull in all dependencies
needed to use the `tox.pytest` module, add a new `test` extra with a
subset of the test dependency group. Though the pytest package should
be present in the users' environment anyway, we still add it for
consistency and to produce a conflict just in case the version has
been constrained to an older version by the project.