Use ubuntu-24.04 during build
This change will ensure that we are in control of the ubuntu version
using build and avoid surprises from GitHub as the date when they
make the switch is not predictable.

This needs to change only once every two years, so not really a
problem. Also sorts the noise we currently get due to the upcoming
switch:

ubuntu-latest pipelines will use ubuntu-24.04 soon. For more details, see https://github.com/actions/runner-images/issues/10636