💅 Make SVG image compatible with Firefox
It appears that Firefox has bugs around interpreting `<feBlend>`
filters in SVG documents. This makes it render the logo as a
transparent image in the docs, both in the mobile and the desktop
versions.

This patch applies a workaround [[1]] found on the internet to make it
work without waiting for Firefox to fix their bug.

Resolves #3465.

[1]: https://github.com/svg/svgo/issues/732